remove analytics which are trivial to compute: gate_volume, topological_swap_count
